✨ Feat: add ability scores component and replace characteristics in player and NPC tabs; introduce new circle icons for death saves
